+++ bignp256/src/arithmetic/scalar.rs
@@ -89,19 +89,6 @@ primefield::monty_field_fiat_arithmetic!(
 
 elliptic_curve::scalar_impls!(BignP256, Scalar);
 
-impl Scalar {
-    /// Returns the square root of self mod p, or `None` if no square root
-    /// exists.
-    pub fn sqrt(&self) -> CtOption<Self> {
-        /// Because p ≡ 3 mod 4, sqrt can be done with only one
-        /// exponentiation via the computation of self^((p + 1) // 4) (mod p).
-        const EXP: U256 =
-            U256::from_be_hex("3ffffffffffffffffffffffffffffffff65723b5837ed37f1f96afe6498f5982");
-        let sqrt = self.pow_vartime(&EXP);
-        CtOption::new(sqrt, (sqrt * sqrt).ct_eq(self))
-    }
-}
-
 impl AsRef<Scalar> for Scalar {
     fn as_ref(&self) -> &Scalar {
         self
