+++ bp256/src/arithmetic/scalar.rs
@@ -72,22 +72,6 @@ primefield::monty_field_fiat_arithmetic!(
 elliptic_curve::scalar_impls!(BrainpoolP256r1, Scalar);
 elliptic_curve::scalar_impls!(BrainpoolP256t1, Scalar);
 
-impl Scalar {
-    /// Returns the square root of self mod n, or `None` if no square root
-    /// exists.
-    pub fn sqrt(&self) -> CtOption<Self> {
-        /// Because n ≡ 3 mod 4 for brainpoolP256's scalar field modulus, sqrt
-        /// can be implemented with only one exponentiation via the computation
-        /// of self^((n + 1) // 4) (mod n).
-        const EXP: U256 =
-            U256::from_be_hex("2a7ed5f6e87baa6f0f9982a42760e35c630e5ea8ed5869bde40783a0a5d215aa");
-
-        // Note: vartime only with respect to `EXP`
-        let sqrt = self.pow_vartime(&EXP);
-        CtOption::new(sqrt, sqrt.square().ct_eq(self))
-    }
-}
-
 impl AsRef<Scalar> for Scalar {
     fn as_ref(&self) -> &Scalar {
         self
