+++ bp384/src/arithmetic/scalar.rs
@@ -72,25 +72,6 @@ primefield::monty_field_fiat_arithmetic!(
 elliptic_curve::scalar_impls!(BrainpoolP384r1, Scalar);
 elliptic_curve::scalar_impls!(BrainpoolP384t1, Scalar);
 
-impl Scalar {
-    /// Atkin algorithm for q mod 8 = 5
-    /// <https://eips.ethereum.org/assets/eip-3068/2012-685_Square_Root_Even_Ext.pdf>
-    /// (page 10, algorithm 3)
-    pub fn sqrt(&self) -> CtOption<Self> {
-        const EXP: U384 = U384::from_be_hex(
-            "119723d054670da501ebadefca1cc83be2a5ee213daa8ad663e2cdcd958084b4f9e756d5ed6ff862077106405d208cac",
-        );
-        let t = Self::from_u64(2).pow_vartime(&EXP);
-        let a1 = self.pow_vartime(&EXP);
-        let a0 = (a1.square() * self).square();
-        let b = t * a1;
-        let ab = self * &b;
-        let i = Self::from_u64(2) * ab * b;
-        let x = ab * (i - Self::ONE);
-        CtOption::new(x, !a0.ct_eq(&-Self::ONE))
-    }
-}
-
 impl AsRef<Scalar> for Scalar {
     fn as_ref(&self) -> &Scalar {
         self
