+++ p192/src/arithmetic/field.rs
@@ -69,20 +69,6 @@ primefield::monty_field_fiat_arithmetic!(
     fiat_p192_selectznz
 );
 
-impl FieldElement {
-    /// Returns the square root of self mod p, or `None` if no square root
-    /// exists.
-    pub fn sqrt(&self) -> CtOption<Self> {
-        /// Because p ≡ 3 mod 4 for secp192r1's base field modulus, sqrt can be done with only one
-        /// exponentiation via the computation of self^((p + 1) // 4) (mod p).
-        const EXP: U192 = U192::from_be_hex("3fffffffffffffffffffffffffffffffc000000000000000");
-
-        // Note: vartime only with respect to `EXP`, which is fixed (so therefore not vartime)
-        let sqrt = self.pow_vartime(&EXP);
-        CtOption::new(sqrt, sqrt.square().ct_eq(self))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::{FieldElement, U192};
